// backend/src/models/QuestionPaper.ts
import mongoose, { Document, Schema } from 'mongoose';

export interface IQuestionPaper extends Document {
  title: string;
  subjectId: mongoose.Types.ObjectId;
  totalMarks: number;
  durationMinutes: number;
  questions: mongoose.Types.ObjectId[];
  createdBy: mongoose.Types.ObjectId;
  status: 'Draft' | 'Published';
  isAutoGenerated: boolean;
}

const questionPaperSchema: Schema = new mongoose.Schema(
  {
    title: { type: String, required: true },
    subjectId: { type: mongoose.Schema.Types.ObjectId, ref: 'Subject', required: true },
    totalMarks: { type: Number, required: true },
    durationMinutes: { type: Number, required: true },
    questions: [{ type: mongoose.Schema.Types.ObjectId, ref: 'Question' }],
    createdBy: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true },
    status: { type: String, enum: ['Draft', 'Published'], default: 'Draft' },
    isAutoGenerated: { type: Boolean, default: false },
  },
  { timestamps: true }
);

const QuestionPaper = mongoose.model<IQuestionPaper>('QuestionPaper', questionPaperSchema);

export default QuestionPaper;